upvote using js request and disable own upvote

+++ app/controllers/upvote_controller.rb
@@ -1,14 +1,26 @@
 class UpvoteController < ApplicationController
 
 	# POST /upvote
-	def newUpvote
-		answer = Answer.find(params[:ans_id])
-		upvote = Upvote.where(users: current_user, answer: answer).first
+	def toggle_upvote
+		@answer = Answer.find(params[:ans_id])
+
+		#prevent user from upvoting own answer
+		if @answer.my_ans? current_user.id
+			return redirect_to '/myAnswers', :flash => { :alert => "Cannot upvote you own answer! :(" }
+		end
+
+		upvote = Upvote.where(users: current_user, answer: @answer).first
+
 		if upvote
 			upvote.destroy!
+			@is_upvoted	= false
 		else
-			Upvote.create(users_id: current_user.id, answer_id: answer.id)
+			Upvote.create(users_id: current_user.id, answer_id: @answer.id)
+			@is_upvoted = true
 		end	
-		return redirect_to '/'
+
+		respond_to do |format|
+			format.js { }
+		end
 	end
 end
